+++ sekizai/context.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from django.template import Context
 
 from sekizai.context_processors import sekizai
@@ -10,5 +9,5 @@ class SekizaiContext(Context):
     no request is available.
     """
     def __init__(self, *args, **kwargs):
-        super(SekizaiContext, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         self.update(sekizai())